+++ sql/samples/core-data.sql
@@ -64,28 +64,28 @@ VALUES
 ('PRO-HLP', 'Housing, Land and Property', '2023-01-01 00:00:00', NULL);
 
 -- dummy data
-INSERT INTO Location (id, code, name, centroid_lat, centroid_lon, reference_period_start, reference_period_end)
+INSERT INTO Location (id, code, name, reference_period_start, reference_period_end)
 VALUES
-(1, 'FOO', 'Foolandia', 1, -1, '2023-01-01 00:00:00', NULL);
+(1, 'FOO', 'Foolandia', '2023-01-01 00:00:00', NULL);
 
 -- dummy data
-INSERT INTO Admin1 (id, location_ref, code, name, centroid_lat, centroid_lon, is_unspecified, reference_period_start, reference_period_end)
+INSERT INTO Admin1 (id, location_ref, code, name, is_unspecified, reference_period_start, reference_period_end)
 VALUES
-(1, 1, 'FOO-XXX', 'Unspecified', 1, -1, TRUE, '2023-01-01 00:00:00', NULL),
-(2, 1, 'FOO-001', 'Province 01', 1, -1, FALSE, '2023-01-01 00:00:00', NULL),
-(3, 1, 'FOO-002', 'Province 02', 1, -1, FALSE, '2023-01-01 00:00:00', NULL);
+(1, 1, 'FOO-XXX', 'Unspecified', TRUE, '2023-01-01 00:00:00', NULL),
+(2, 1, 'FOO-001', 'Province 01', FALSE, '2023-01-01 00:00:00', NULL),
+(3, 1, 'FOO-002', 'Province 02', FALSE, '2023-01-01 00:00:00', NULL);
 
 -- dummy data
 -- note that we need an "Unspecified" for every Admin1, including the unspecified one
-INSERT INTO Admin2 (id, admin1_ref, code, name, centroid_lat, centroid_lon, is_unspecified, reference_period_start, reference_period_end)
+INSERT INTO Admin2 (id, admin1_ref, code, name, is_unspecified, reference_period_start, reference_period_end)
 VALUES
-(1, 1, 'FOO-XXX-XXX', 'Unspecified', 1, -1, TRUE, '2023-01-01 00:00:00', NULL),
-(2, 2, 'FOO-001-XXX', 'Unspecified', 1, -1, TRUE, '2023-01-01 00:00:00', NULL),
-(3, 3, 'FOO-002-XXX', 'Unspecified', 1, -1, TRUE, '2023-01-01 00:00:00', NULL),
-(4, 2, 'FOO-001-A', 'District A', 1, -1, FALSE, '2023-01-01 00:00:00', NULL),
-(5, 2, 'FOO-001-B', 'District B', 1, -1, FALSE, '2023-01-01 00:00:00', NULL),
-(6, 3, 'FOO-002-C', 'District C', 1, -1, FALSE, '2023-01-01 00:00:00', NULL),
-(7, 3, 'FOO-002-D', 'District D', 1, -1, FALSE, '2023-01-01 00:00:00', NULL);
+(1, 1, 'FOO-XXX-XXX', 'Unspecified', TRUE, '2023-01-01 00:00:00', NULL),
+(2, 2, 'FOO-001-XXX', 'Unspecified', TRUE, '2023-01-01 00:00:00', NULL),
+(3, 3, 'FOO-002-XXX', 'Unspecified', TRUE, '2023-01-01 00:00:00', NULL),
+(4, 2, 'FOO-001-A', 'District A', FALSE, '2023-01-01 00:00:00', NULL),
+(5, 2, 'FOO-001-B', 'District B', FALSE, '2023-01-01 00:00:00', NULL),
+(6, 3, 'FOO-002-C', 'District C', FALSE, '2023-01-01 00:00:00', NULL),
+(7, 3, 'FOO-002-D', 'District D', FALSE, '2023-01-01 00:00:00', NULL);
 
 
 -- may be OK for production